fix login modal
